警告を修正。

+++ Game/Game/Player/Player.h
@@ -62,7 +62,7 @@ namespace nsAWA {
 				//SPを回復。
 				m_status.HealSP(healValue);
 			}
-			void AddExp(float exp) {
+			void AddExp(int exp) {
 
 				//経験値を獲得。
 				m_status.AddExp(exp);
